tests: test post & put methods
