+++ pkg/checks/checks.go
@@ -74,6 +74,10 @@ type Result struct {
 	Err string `json:"error"`
 }
 
+type GlobalTarget struct {
+	Url      string    `json:"url"`
+	LastSeen time.Time `json:"lastSeen"`
+}
 type ResultDTO struct {
 	Name   string
 	Result *Result
